订单同步接口修改

+++ order-core/src/main/java/com/ai/slp/order/service/business/interfaces/ISyncronizeBusiSV.java
@@ -13,6 +13,6 @@ public interface ISyncronizeBusiSV {
 	 * @throws BusinessException
 	 * @throws SystemException
 	 */
-	public int orderSynchronize(OrderSynchronizeVo request) throws BusinessException, SystemException;
+	public long orderSynchronize(OrderSynchronizeVo request) throws BusinessException, SystemException;
 
 }
